+++ packages/polaris-viz-core/src/hooks/useSparkBar.ts
@@ -115,9 +115,6 @@ export function useSparkBar({
         },
       ];
 
-  const radius = Math.floor(barWidth / 2);
-  const borderRadius = `${radius} ${radius} 0 0`;
-
   return {
     dataWithIndex,
     color,
@@ -127,7 +124,6 @@ export function useSparkBar({
     getBarHeight,
     xScale,
     yScale,
-    borderRadius,
     targetLineYPosition,
     targetLineX1: 0 - offsetLeft,
     targetLineX2: width + offsetRight,