+++ packages/polaris-viz-core/src/types.ts
@@ -76,8 +76,8 @@ export interface ArcTheme {
 }
 
 export interface BarTheme {
-  hasRoundedCorners: boolean;
   zeroValueColor: string;
+  borderRadius: number;
 }
 
 export interface XAxisTheme {